+++ src/main/java/raisetech/student/management/controller/StudentController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import raisetech.student.management.domain.StudentDetail;
-import raisetech.student.management.domain.validation.CreateGroup;
-import raisetech.student.management.domain.validation.UpdateGroup;
+import raisetech.student.management.validation.CreateGroup;
+import raisetech.student.management.validation.UpdateGroup;
 import raisetech.student.management.service.StudentService;
 
 /**
