+++ src/main/java/raisetech/student/management/validation/UpdateGroup.java
@@ -1,4 +1,4 @@
-package raisetech.student.management.domain.validation;
+package raisetech.student.management.validation;
 
 /**
  * 更新時のバリデーション用グループ
